:recycle: refactor stores, fix wallet
