+++ package.json
@@ -30,7 +30,7 @@
     "@css-render/vue3-ssr": "^0.15.11",
     "@kodadot1/sub-api": "0.1.1-alpha.3",
     "@paraspell/sdk": "^0.0.16",
-    "@pinia/nuxt": "^0.4.3",
+    "@pinia/nuxt": "^0.4.6",
     "@polkadot/api": "^9.6.2",
     "@polkadot/extension-dapp": "^0.44.6",
     "@polkadot/extension-inject": "^0.44.6",
