+++ plugins/paraspell.ts
@@ -1,0 +1,9 @@
+import * as paraspell from '@paraspell/sdk'
+
+export default defineNuxtPlugin(() => {
+  return {
+    provide: {
+      paraspell,
+    },
+  }
+})
